+++ sdk/dotnet/GetRegion.cs
@@ -15,7 +15,7 @@ namespace Pulumi.Linode
         /// &gt; This content is derived from https://github.com/terraform-providers/terraform-provider-linode/blob/master/website/docs/d/region.html.markdown.
         /// </summary>
         public static Task<GetRegionResult> GetRegion(GetRegionArgs args, InvokeOptions? options = null)
-            => Pulumi.Deployment.Instance.InvokeAsync<GetRegionResult>("linode:index/getRegion:getRegion", args, options.WithVersion());
+            => Pulumi.Deployment.Instance.InvokeAsync<GetRegionResult>("linode:index/getRegion:getRegion", args ?? ResourceArgs.Empty, options.WithVersion());
     }
 
     public sealed class GetRegionArgs : Pulumi.ResourceArgs