+++ sdk/dotnet/GetUser.cs
@@ -25,7 +25,7 @@ namespace Pulumi.Linode
         /// &gt; This content is derived from https://github.com/terraform-providers/terraform-provider-linode/blob/master/website/docs/d/user.html.markdown.
         /// </summary>
         public static Task<GetUserResult> GetUser(GetUserArgs args, InvokeOptions? options = null)
-            => Pulumi.Deployment.Instance.InvokeAsync<GetUserResult>("linode:index/getUser:getUser", args, options.WithVersion());
+            => Pulumi.Deployment.Instance.InvokeAsync<GetUserResult>("linode:index/getUser:getUser", args ?? ResourceArgs.Empty, options.WithVersion());
     }
 
     public sealed class GetUserArgs : Pulumi.ResourceArgs