+++ sdk/dotnet/Token.cs
@@ -51,7 +51,7 @@ namespace Pulumi.Linode
         /// <param name="args">The arguments used to populate this resource's properties</param>
         /// <param name="options">A bag of options that control this resource's behavior</param>
         public Token(string name, TokenArgs args, CustomResourceOptions? options = null)
-            : base("linode:index/token:Token", name, args, MakeResourceOptions(options, ""))
+            : base("linode:index/token:Token", name, args ?? ResourceArgs.Empty, MakeResourceOptions(options, ""))
         {
         }