+++ sdk/dotnet/Volume.cs
@@ -76,7 +76,7 @@ namespace Pulumi.Linode
         /// <param name="args">The arguments used to populate this resource's properties</param>
         /// <param name="options">A bag of options that control this resource's behavior</param>
         public Volume(string name, VolumeArgs args, CustomResourceOptions? options = null)
-            : base("linode:index/volume:Volume", name, args, MakeResourceOptions(options, ""))
+            : base("linode:index/volume:Volume", name, args ?? ResourceArgs.Empty, MakeResourceOptions(options, ""))
         {
         }