+++ sdk/python/pulumi_linode/object_storage_key.py
@@ -9,18 +9,30 @@ import pulumi.runtime
 from typing import Union
 from . import utilities, tables
 
-class Provider(pulumi.ProviderResource):
-    def __init__(__self__, resource_name, opts=None, api_version=None, token=None, ua_prefix=None, url=None, __props__=None, __name__=None, __opts__=None):
+class ObjectStorageKey(pulumi.CustomResource):
+    access_key: pulumi.Output[str]
+    label: pulumi.Output[str]
+    """
+    The label given to this key. For display purposes only.
+    """
+    secret_key: pulumi.Output[str]
+    def __init__(__self__, resource_name, opts=None, label=None, __props__=None, __name__=None, __opts__=None):
         """
-        The provider type for the linode package. By default, resources use package-wide configuration
-        settings, however an explicit `Provider` instance may be created and passed during resource
-        construction to achieve fine-grained programmatic control over provider settings. See the
-        [documentation](https://www.pulumi.com/docs/reference/programming-model/#providers) for more information.
+        Provides a Linode Object Storage Key resource. This can be used to create, modify, and delete Linodes Object Storage Keys.
+        
+        ## Attributes
+        
+        This resource exports the following attributes:
+        
+        * `access_key` - This keypair's access key. This is not secret.
+        
+        * `secret_key` - This keypair's secret key.
         
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
+        :param pulumi.Input[str] label: The label given to this key. For display purposes only.
 
-        > This content is derived from https://github.com/terraform-providers/terraform-provider-linode/blob/master/website/docs/index.html.markdown.
+        > This content is derived from https://github.com/terraform-providers/terraform-provider-linode/blob/master/website/docs/r/object_storage_key.html.markdown.
         """
         if __name__ is not None:
             warnings.warn("explicit use of __name__ is deprecated", DeprecationWarning)
@@ -39,34 +51,37 @@ class Provider(pulumi.ProviderResource):
                 raise TypeError('__props__ is only valid when passed in combination with a valid opts.id to get an existing resource')
             __props__ = dict()
 
-            __props__['api_version'] = api_version
-            if token is None:
-                token = utilities.get_env('LINODE_TOKEN', 'LINODE_API_TOKEN')
-            __props__['token'] = token
-            __props__['ua_prefix'] = ua_prefix
-            __props__['url'] = url
-        super(Provider, __self__).__init__(
-            'linode',
+            if label is None:
+                raise TypeError("Missing required property 'label'")
+            __props__['label'] = label
+            __props__['access_key'] = None
+            __props__['secret_key'] = None
+        super(ObjectStorageKey, __self__).__init__(
+            'linode:index/objectStorageKey:ObjectStorageKey',
             resource_name,
             __props__,
             opts)
 
     @staticmethod
-    def get(resource_name, id, opts=None):
+    def get(resource_name, id, opts=None, access_key=None, label=None, secret_key=None):
         """
-        Get an existing Provider resource's state with the given name, id, and optional extra
+        Get an existing ObjectStorageKey resource's state with the given name, id, and optional extra
         properties used to qualify the lookup.
         
         :param str resource_name: The unique name of the resulting resource.
         :param str id: The unique provider ID of the resource to lookup.
         :param pulumi.ResourceOptions opts: Options for the resource.
+        :param pulumi.Input[str] label: The label given to this key. For display purposes only.
 
-        > This content is derived from https://github.com/terraform-providers/terraform-provider-linode/blob/master/website/docs/index.html.markdown.
+        > This content is derived from https://github.com/terraform-providers/terraform-provider-linode/blob/master/website/docs/r/object_storage_key.html.markdown.
         """
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(id=id))
 
         __props__ = dict()
-        return Provider(resource_name, opts=opts, __props__=__props__)
+        __props__["access_key"] = access_key
+        __props__["label"] = label
+        __props__["secret_key"] = secret_key
+        return ObjectStorageKey(resource_name, opts=opts, __props__=__props__)
     def translate_output_property(self, prop):
         return tables._CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop