+++ CHANGELOG.md
@@ -3,6 +3,8 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 * Upgrade to support go 1.13.x
+* Upgrade to pulumi-terraform-bridge v1.4.2
+* Upgrade to v1.9.0 of the Linode Terraform Provider
 
 ---